chore: Renamame withdrawal components (#3396)
* chore: Renamame withdrawal components

* Fix component import

---------
