+++ src/components/forms/pool_actions/WithdrawForm/WithdrawForm.vue
@@ -1,13 +1,12 @@
 <script setup lang="ts">
-import { onBeforeMount, computed } from 'vue';
 import TokenInput from '@/components/inputs/TokenInput/TokenInput.vue';
 import { isLessThanOrEqualTo, isRequired } from '@/lib/utils/validations';
 import useWeb3 from '@/services/web3/useWeb3';
-import ProportionalWithdrawalInputV2 from './components/ProportionalWithdrawalInputV2.vue';
-import WithdrawTotalsV2 from './components/WithdrawTotalsV2.vue';
+import ProportionalWithdrawalInput from './components/ProportionalWithdrawalInput.vue';
+import WithdrawTotals from './components/WithdrawTotals.vue';
 import { useExitPool } from '@/providers/local/exit-pool.provider';
 import useVeBal from '@/composables/useVeBAL';
-import WithdrawPreviewModalV2 from './components/WithdrawPreviewModal/WithdrawPreviewModalV2.vue';
+import WithdrawPreviewModal from './components/WithdrawPreviewModal/WithdrawPreviewModal.vue';
 import { useTokens } from '@/providers/tokens.provider';
 import {
   isPreMintedBptType,
@@ -97,7 +96,7 @@ onBeforeMount(() => {
 
 <template>
   <div data-testid="withdraw-form">
-    <ProportionalWithdrawalInputV2 v-if="!isSingleAssetExit" :pool="pool" />
+    <ProportionalWithdrawalInput v-if="!isSingleAssetExit" :pool="pool" />
     <template v-else>
       <!-- Single asset exit input -->
       <TokenInput
@@ -116,7 +115,7 @@ onBeforeMount(() => {
       />
     </template>
 
-    <WithdrawTotalsV2 class="mt-4" />
+    <WithdrawTotals class="mt-4" />
 
     <div
       v-if="highPriceImpact"
@@ -165,7 +164,7 @@ onBeforeMount(() => {
     </div>
 
     <teleport to="#modal">
-      <WithdrawPreviewModalV2
+      <WithdrawPreviewModal
         v-if="showPreview"
         :pool="pool"
         @close="showPreview = false"
