+++ src/components/forms/pool_actions/WithdrawForm/components/ProportionalWithdrawalTokenInfo.vue
