+++ src/components/forms/pool_actions/WithdrawForm/components/WithdrawTotals.vue
@@ -1,6 +1,4 @@
 <script setup lang="ts">
-import { computed } from 'vue';
-
 import useNumbers, { FNumFormats } from '@/composables/useNumbers';
 
 import { useExitPool } from '@/providers/local/exit-pool.provider';
